+++ extensions/Worker.Extensions.ServiceBus/src/Constants.cs
@@ -1,0 +1,11 @@
+// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the MIT License. See License.txt in the project root for license information.
+
+
+namespace Microsoft.Azure.Functions.Worker.Extensions.ServiceBus
+{
+    public static class Constants
+    {
+        internal const string BinaryContentType = "application/octet-stream";
+    }
+}