+++ sdk/Sdk/FunctionMetadataJsonWriter.cs
@@ -1,4 +1,4 @@
-// Copyright (c) .NET Foundation. All rights reserved.
+﻿// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
 
 using System.Collections.Generic;
@@ -10,39 +10,33 @@ namespace Microsoft.Azure.Functions.Worker.Sdk
 {
     internal static class FunctionMetadataJsonWriter
     {
-        private static readonly JsonSerializerOptions _serializerOptions = CreateSerializerOptions();
-        private const string _fileName = "functions.metadata";
+        private const string FileName = "functions.metadata";
+        private static readonly JsonSerializerOptions s_serializerOptions = CreateSerializerOptions();
 
         private static JsonSerializerOptions CreateSerializerOptions()
         {
             var namingPolicy = new FunctionsJsonNamingPolicy();
-
-            var options = new JsonSerializerOptions
+            return new JsonSerializerOptions
             {
                 WriteIndented = true,
                 IgnoreNullValues = true,
                 PropertyNameCaseInsensitive = true,
                 IgnoreReadOnlyProperties = true,
                 DictionaryKeyPolicy = namingPolicy,
-                PropertyNamingPolicy = namingPolicy
+                PropertyNamingPolicy = namingPolicy,
+                Converters =
+                {
+                    new JsonStringEnumConverter(),
+                }
             };
-
-            options.Converters.Add(new JsonStringEnumConverter());
-
-            return options;
         }
 
         public static void WriteMetadata(IEnumerable<SdkFunctionMetadata> functions, string metadataFileDirectory)
         {
-            string metadataFile = Path.Combine(metadataFileDirectory, _fileName);
-
-            using (var fs = new FileStream(metadataFile, FileMode.Create, FileAccess.Write))
-            {
-                using (var writer = new Utf8JsonWriter(fs, new JsonWriterOptions { Indented = true }))
-                {
-                    JsonSerializer.Serialize(writer, functions, _serializerOptions);
-                }
-            }
+            string metadataFile = Path.Combine(metadataFileDirectory, FileName);
+            using var fs = new FileStream(metadataFile, FileMode.Create, FileAccess.Write);
+            using var writer = new Utf8JsonWriter(fs, new JsonWriterOptions { Indented = true });
+            JsonSerializer.Serialize(writer, functions, s_serializerOptions);
         }
     }
 }