+++ extensions/Worker.Extensions.Storage/release_notes.md
@@ -1,4 +1,17 @@
-## Release notes
+## What's Changed
+
 <!-- Please add your release notes in the following format:
 - My change description (#PR/#issue)
 -->
+
+### Microsoft.Azure.Functions.Worker.Extensions.Storage <version> (metapackage)
+
+- <entry>
+
+### Microsoft.Azure.Functions.Worker.Extensions.Storage.Blobs <version> (delete if not updated)
+
+- <entry>
+
+### Microsoft.Azure.Functions.Worker.Extensions.Storage.Queues <version> (delete if not updated)
+
+- <entry>